Move all legacy handling in `but-ctx` into its own module.
